@@ -70,25 +79,28 @@
     "karma-electron": "6.3.1",
     "karma-tap": "4.2.0",
     "karma-webpack": "4.0.2",
-    "lerna": "3.20.2",
+    "lerna": "3.22.1",
     "lint-staged": "10.2.2",
     "npm-run-all": "4.1.5",
     "prettier": "2.0.5",
     "secp256k1": "4.0.0",
+    "shebang-loader": "0.0.1",
     "source-map-loader": "0.2.4",
+    "stream-browserify": "3.0.0",
     "tap": "14.10.6",
     "tape": "5.0.1",
+    "tape-promise": "4.0.0",
     "ts-loader": "6.2.1",
     "ts-node": "8.9.1",
     "tslint": "6.0.0",
     "typescript": "4.0.3",
-    "webpack": "4.41.6",
+    "webpack": "git+https://github.com/petermetz/webpack.git#a1021fd674dd24b13d759611b77b5d7057f7f0ea",
     "webpack-bundle-analyzer": "3.6.0",
     "webpack-cli": "3.3.11"
   },
   "config": {
     "commitizen": {
-      "path": "./node_modules/cz-conventional-changelog"
+      "path": "./node_modules/git-cz"
     }
   },
   "husky": {
