@@ -11,7 +11,7 @@ import {
 } from "@hyperledger/cactus-common";
 import { FORMAT_PLUGIN_ARRAY } from "./convict-plugin-array-format";
 import { SelfSignedPkiGenerator, IPki } from "./self-signed-pki-generator";
-import { Consortium } from "@hyperledger/cactus-plugin-consortium-manual";
+import { Consortium } from "@hyperledger/cactus-core-api";
 
 convict.addFormat(FORMAT_PLUGIN_ARRAY);
 convict.addFormat(ipaddress);
@@ -410,16 +410,22 @@ export class ConfigService {
 
     const keyPair = JWK.generateSync("EC", "secp256k1", { use: "sig" }, true);
     const keyPairPem = keyPair.toPEM(true);
+    const memberId1 = "Cactus_Example_Consortium_Member_1";
     const consortium: Consortium = {
       name: "Example Cactus Consortium",
       id: uuidV4(),
       mainApiHost: apiBaseUrl,
       members: [
         {
-          id: uuidV4(),
+          id: memberId1,
           name: "Example Cactus Consortium Member 1",
           nodes: [
             {
+              consortiumId: "Cactus_Example_Consortium",
+              id: "Cactus_Example_Consortium_Node_1",
+              ledgers: [],
+              memberId: memberId1,
+              plugins: [],
               nodeApiHost: apiBaseUrl,
               publicKeyPem: keyPair.toPEM(false),
             },
