@@ -1,3 +1,4 @@
+export * from "./generated/openapi/typescript-axios/index";
 export { IPluginLedgerConnector } from "./plugin/ledger-connector/i-plugin-ledger-connector";
 export { IPluginKVStorage } from "./plugin/storage/key-value/i-plugin-kv-storage";
 export { IPluginKeychain } from "./plugin/keychain/i-plugin-keychain";
@@ -8,6 +9,10 @@ export {
 } from "./plugin/web-service/i-plugin-web-service";
 export { IWebServiceEndpoint } from "./plugin/web-service/i-web-service-endpoint";
 export { PluginFactory } from "./plugin/plugin-factory";
-export { ICactusPlugin, isICactusPlugin } from "./plugin/i-cactus-plugin";
+export {
+  ICactusPlugin,
+  ICactusPluginOptions,
+  isICactusPlugin,
+} from "./plugin/i-cactus-plugin";
 export { PluginAspect } from "./plugin/plugin-aspect";
 export { PluginRegistry } from "./plugin/plugin-registry";
