@@ -1,26 +1,30 @@
-import { Request, Response, NextFunction } from "express";
-import { JWS, JWK } from "jose";
+import { Express, Request, Response, NextFunction } from "express";
 import flatten from "lodash/flatten";
+import { AxiosResponse } from "axios";
 
-import { ApiClient } from "@hyperledger/cactus-api-client";
 import {
   IWebServiceEndpoint,
   IExpressRequestHandler,
-} from "@hyperledger/cactus-core-api";
-import {
-  Configuration,
-  DefaultApi,
   JWSGeneral,
   JWSRecipient,
   Consortium,
+} from "@hyperledger/cactus-core-api";
+
+import {
+  Configuration,
+  DefaultApi,
   GetNodeJwsResponse,
 } from "../generated/openapi/typescript-axios";
+
 import {
   LogLevelDesc,
   Logger,
   LoggerProvider,
 } from "@hyperledger/cactus-common";
-import { AxiosResponse } from "axios";
+
+import { registerWebServiceEndpoint } from "@hyperledger/cactus-core";
+
+import { GetConsortiumEndpointV1 as Constants } from "./get-consortium-jws-endpoint-constants";
 
 export interface IGetConsortiumJwsEndpointOptions {
   keyPairPem: string;
@@ -57,10 +61,23 @@ export class GetConsortiumEndpointV1 implements IWebServiceEndpoint {
   }
 
   getPath(): string {
-    return this.options.path;
+    return Constants.HTTP_PATH;
   }
 
-  async handleRequest(req: Request, res: Response): Promise<void> {
+  getVerbLowerCase(): string {
+    return Constants.HTTP_VERB_LOWER_CASE;
+  }
+
+  registerExpress(app: Express): IWebServiceEndpoint {
+    registerWebServiceEndpoint(app, this);
+    return this;
+  }
+
+  async handleRequest(
+    req: Request,
+    res: Response,
+    next: NextFunction
+  ): Promise<void> {
     const fnTag = "GetConsortiumJwsEndpointV1#handleRequest()";
     this.log.debug(`GET ${this.getPath()}`);
 
@@ -71,8 +88,7 @@ export class GetConsortiumEndpointV1 implements IWebServiceEndpoint {
       const requests: Promise<AxiosResponse<GetNodeJwsResponse>>[] = nodes
         .map((cnm) => cnm.nodeApiHost)
         .map((host) => new Configuration({ basePath: host }))
-        .map((configuration) => new ApiClient(configuration))
-        .map((apiClient) => apiClient.extendWith(DefaultApi))
+        .map((configuration) => new DefaultApi(configuration))
         .map((apiClient) =>
           apiClient.apiV1PluginsHyperledgerCactusPluginConsortiumManualNodeJwsGet()
         );
