@@ -1,15 +1,22 @@
+import { v4 as uuidv4 } from "uuid";
+
 import { PluginFactory } from "@hyperledger/cactus-core-api";
 import {
-  IPluginKeychainOptions,
+  IPluginKeychainMemoryOptions,
   PluginKeychainMemory,
 } from "./plugin-keychain-memory";
 
 export class PluginFactoryKeychain extends PluginFactory<
   PluginKeychainMemory,
-  IPluginKeychainOptions
+  IPluginKeychainMemoryOptions
 > {
   async create(
-    options: IPluginKeychainOptions = { backend: new Map() }
+    options: IPluginKeychainMemoryOptions = {
+      backend: new Map(),
+      instanceId: uuidv4(),
+      keychainId: uuidv4(),
+      logLevel: "TRACE",
+    }
   ): Promise<PluginKeychainMemory> {
     return new PluginKeychainMemory(options);
   }
