@@ -1,11 +1,13 @@
+export * from "./generated/openapi/typescript-axios";
+
 export {
   PluginLedgerConnectorQuorum,
   IPluginLedgerConnectorQuorumOptions,
-  IQuorumDeployContractOptions,
-  ITransactionOptions,
 } from "./plugin-ledger-connector-quorum";
+
+export * from "./model-type-guards";
+
 export { PluginFactoryLedgerConnector } from "./plugin-factory-ledger-connector";
-export { Contract as Web3EthContract } from "web3-eth-contract";
 
 import { PluginFactoryLedgerConnector } from "./plugin-factory-ledger-connector";
 export async function createPluginFactory(
